+++ lib/client/app.ml
@@ -34,6 +34,7 @@ let app_computation =
     | Home -> Pages.Page_home.component ~theme ()
     | Work -> Pages.Page_work.component ~theme ()
     | About -> Pages.Page_about.component ~theme ()
+    | Me -> Pages.Page_me.component ~theme ()
     | Projects -> Pages.Page_projects.component ~theme ()
     | Project_detail slug -> Pages.Page_project_detail.component ~theme ~slug ()
     | Resume -> Pages.Page_resume.component ~theme ()